Move init_test
